refactor(dashboard-api): extract pure settings helpers into settings.py
